pyproject.toml
[project]
name = "timm-explorer"
version = "0.1.0"
description = "A Taipy web application for exploring TIMM models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "taipy-gui>=3.0.0",
    "timm==0.9.12",
    "pandas==2.2.0",
    "torch==2.2.0",
    "numpy==1.26.4",
    "pyarrow>=15.0.0",
    "tqdm>=4.66.0",
    "fastparquet>=2024.11.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
]
